Fix article comment handling. Resolves #49. More histotical context in the comment within that issue.
Additional message information fetching mechanism was implemented at the
time KockaLogger was using no `async`/`await`, so everything was
callbacks and things were being passed through `this`. As this is no
longer the case, our message fetching mechanism had a bit of refactoring
here.

In addition to that, to actually fix the linked issue,
`DiscussionsMessage#fetch` was implemented, similarly to how we fetched
page titles with Fandom's old thread system.